Read only mode

+++ Src/LastRun/LastRun.py
@@ -2,25 +2,32 @@ from pathlib import Path
 import os
 import datetime
 import Config.ObsidianFilePath as ObsidianFilePath
+import Config.Setting as Setting
 
 LastRunIsToday = False
 
 _lastRunFilePath = Path(ObsidianFilePath.lastRun)
 
-# Last fun file exist
-if (_lastRunFilePath.is_file()):
-    # Get last modified date
-    fileTimeStamp = os.path.getmtime(_lastRunFilePath)
-    fileTime = datetime.datetime.fromtimestamp(fileTimeStamp)
+# When read only
+# skip update last run file and assume it has already updated by server
+if Setting.readOnly:
+    LastRunIsToday = True
 
-    # If today
-    LastRunIsToday = fileTime.date() == datetime.datetime.today().date()
+else:
+    # Last fun file exist
+    if (_lastRunFilePath.is_file()):
+        # Get last modified date
+        fileTimeStamp = os.path.getmtime(_lastRunFilePath)
+        fileTime = datetime.datetime.fromtimestamp(fileTimeStamp)
 
-    if (LastRunIsToday == False):
-        # Update last run file
+        # If today
+        LastRunIsToday = fileTime.date() == datetime.datetime.today().date()
+
+        if (LastRunIsToday == False):
+            # Update last run file
+            with open(_lastRunFilePath, 'w') as emptyFile:
+                pass
+    else:
+        # Create last fun file
         with open(_lastRunFilePath, 'w') as emptyFile:
             pass
-else:
-    # Create last fun file
-    with open(_lastRunFilePath, 'w') as emptyFile:
-        pass
